docs: atomic-icon
