+++ packages/atomic/tsconfig.json
@@ -17,9 +17,6 @@
   },
   "references": [
     {
-      "path": "tsconfig.stories.json"
-    },
-    {
       "path": "./tsconfig.storybook.json"
     }
   ],
