--- packages/atomic/tsconfig.stories.json
@@ -1,24 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "jsx": "react",
-    "plugins": [
-      {
-        "name": "ts-lit-plugin",
-        "strict": true,
-        "rules": {
-          "no-missing-import": "off",
-          "no-unknown-attribute": "off"
-        }
-      }
-    ],
-    "baseUrl": "./",
-    "paths": {
-      "#atomic-storybook/*": ["./storybookUtils/*"]
-    }
-  },
-  "include": ["**/*.stories.tsx", "./storybookUtils/**/*.tsx"]
-}
